Don't throw an exception when merging agents.
The code in ArchivesSpace that updates modification times across
relationship links (to trigger reindexing of graphs of records)
mistakenly assumes that all models will include the Relationship
mixin, and will therefore have an `update_toplevel_mtimes` method.
This is true in the ArchivesSpace core, but not in this plugin.

Add the method to avoid an error in ArchivesSpace 2.0.0/2.0.1.  I'll
correct the issue in the core code too.